requirements.txt
Flask
requests
pandas
beautifulsoup4
python-dotenv
gunicorn
